requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ratelimit==4.1.0
djangorestframework==3.15.2
PyJWT==2.8.0
sqlparse==0.5.1
